Replacing "\tikzname\ " by TikZ in a comment line
As the code is typeset verbatim (all in italic after the % sign), the \tikzname is not rendered as "Ti*k*Z", but as "\tikzname". So I suggest replacing \tikzname by TikZ in comments of code. See page 14 of the current documentation the listing of the code: we read "\tikzname" in the comment, not "TikZ".